authRoute and controller created
