+++ routes/authRoute.js
@@ -1,0 +1,5 @@
+const router = require("express").Router();
+
+router.route("/").post();
+
+module.exports = router;
